Move bundle.py to __init__.py in metadata_api
