+++ src/azul/project/hca/transformers.py
@@ -2,26 +2,16 @@ import logging
 from collections import defaultdict
 from typing import Any, List, MutableMapping, Sequence
 
-from azul.transformer import Document, ElasticSearchDocument, Transformer
-from .metadata_api.bundle import (Bundle,
-                                  EntityVisitor,
-                                  Process,
-                                  LibraryPreparationProcess,
-                                  SequencingProcess,
-                                  Entity,
-                                  Biomaterial,
-                                  UUID4,
-                                  File,
-                                  Protocol,
-                                  SpecimenFromOrganism,
-                                  DonorOrganism,
-                                  SequenceFile, CellSuspension)
 from azul.dss_bundle import DSSBundle
+from azul.transformer import Document, ElasticSearchDocument, Transformer
+
+from azul.project.hca import metadata_api as api
+
 
 log = logging.getLogger(__name__)
 
 
-def _project_dict(bundle: Bundle) -> dict:
+def _project_dict(bundle: api.Bundle) -> dict:
     project = bundle.project
     project = dict(project_shortname=project.project_shortname,
                    laboratory=list(project.laboratory_names),
@@ -31,10 +21,10 @@ def _project_dict(bundle: Bundle) -> dict:
 
 
 def _specimen_dict(
-        biomaterials: MutableMapping[UUID4, Biomaterial]
+        biomaterials: MutableMapping[api.UUID4, api.Biomaterial]
 ) -> List[MutableMapping[str, Any]]:
     specimen_list = []
-    for specimen in (s for s in biomaterials.values() if isinstance(s, SpecimenFromOrganism)):
+    for specimen in (s for s in biomaterials.values() if isinstance(s, api.SpecimenFromOrganism)):
         bio_visitor = BiomaterialVisitor()
         specimen.accept(bio_visitor)
         specimen.ancestors(bio_visitor)
@@ -54,27 +44,27 @@ def _specimen_dict(
     return specimen_list
 
 
-class TransformerVisitor(EntityVisitor):
-    specimens: MutableMapping[UUID4, SpecimenFromOrganism]
+class TransformerVisitor(api.EntityVisitor):
+    specimens: MutableMapping[api.UUID4, api.SpecimenFromOrganism]
     processes: MutableMapping[str, Any]  # Merges process with protocol
-    files: MutableMapping[UUID4, Any]  # Merges manifest + file metadata
+    files: MutableMapping[api.UUID4, Any]  # Merges manifest + file metadata
 
     @staticmethod
-    def _merge_process_protocol(pc: Process, pl: Protocol) -> MutableMapping[str, Any]:
+    def _merge_process_protocol(pc: api.Process, pl: api.Protocol) -> MutableMapping[str, Any]:
         d = dict(documet_id=[pc.document_id, pl.document_id],
                  process_id=pc.process_id,
                  process_name=pc.process_name,
                  protocol_id=pl.protocol_id,
                  protocol_name=pl.protocol_name,
                  _type="process")
-        if isinstance(pc, LibraryPreparationProcess):
+        if isinstance(pc, api.LibraryPreparationProcess):
             d["library_construction_approach"] = pc.library_construction_approach
-        elif isinstance(pc, SequencingProcess):
+        elif isinstance(pc, api.SequencingProcess):
             d["instrument_manufacturer_model"] = pc.instrument_manufacturer_model
         return d
 
     @staticmethod
-    def _file_dict(f: File) -> MutableMapping[str, Any]:
+    def _file_dict(f: api.File) -> MutableMapping[str, Any]:
         d = {
             "content-type": f.manifest_entry.content_type,
             "crc32c": f.manifest_entry.crc32c,
@@ -90,7 +80,7 @@ class TransformerVisitor(EntityVisitor):
             "file_format": f.file_format,
             "_type": "file"
         }
-        if isinstance(f, SequenceFile):
+        if isinstance(f, api.SequenceFile):
             d = {
                 **d,
                 "read_index": f.read_index,
@@ -103,39 +93,39 @@ class TransformerVisitor(EntityVisitor):
         self.processes = {}
         self.files = {}
 
-    def visit(self, entity: Entity) -> None:
-        if isinstance(entity, SpecimenFromOrganism):
+    def visit(self, entity: api.Entity) -> None:
+        if isinstance(entity, api.SpecimenFromOrganism):
             self.specimens[entity.document_id] = entity
-        elif isinstance(entity, Process):
+        elif isinstance(entity, api.Process):
             for pl in entity.protocols.values():
                 pl_pr_id = f"{str(pl.document_id)}-{str(entity.document_id)}"
                 self.processes[pl_pr_id] = self._merge_process_protocol(entity, pl)
-        elif isinstance(entity, File):
+        elif isinstance(entity, api.File):
             self.files[entity.document_id] = self._file_dict(entity)
 
 
-class BiomaterialVisitor(EntityVisitor):
+class BiomaterialVisitor(api.EntityVisitor):
     specimen: MutableMapping[str, Any]
     biomaterial_lineage: List[MutableMapping[str, Any]]
 
     def __init__(self) -> None:
         self.biomaterial_lineage = []
 
-    def visit(self, entity: Entity) -> None:
+    def visit(self, entity: api.Entity) -> None:
 
-        if isinstance(entity, Biomaterial):
+        if isinstance(entity, api.Biomaterial):
             # noinspection PyProtectedMember
             b = dict(document_id=entity.document_id,
                      has_input_biomaterial=entity.has_input_biomaterial,
                      _source=entity._source)
-            if isinstance(entity, SpecimenFromOrganism):
+            if isinstance(entity, api.SpecimenFromOrganism):
                 b["biomaterial_id"] = entity.biomaterial_id
                 b["disease"] = list(entity.disease)
                 b["organ"] = entity.organ
                 b["organ_part"] = entity.organ_part
                 b["storage_method"] = entity.storage_method
                 b["_type"] = "specimen"
-            elif isinstance(entity, DonorOrganism):
+            elif isinstance(entity, api.DonorOrganism):
                 b["donor_biomaterial_id"] = entity.biomaterial_id
                 b["genus_species"] = entity.genus_species
                 b["disease"] = list(entity.disease)
@@ -144,7 +134,7 @@ class BiomaterialVisitor(EntityVisitor):
                 b["max_organism_age_in_seconds"] = entity.organism_age_in_seconds.max
                 b["min_organism_age_in_seconds"] = entity.organism_age_in_seconds.min
                 b["biological_sex"] = entity.biological_sex
-            elif isinstance(entity, CellSuspension):
+            elif isinstance(entity, api.CellSuspension):
                 b["total_estimated_cells"] = entity.total_estimated_cells
             # As more facets are required by the browser, handle each biomateiral as appropriate
             self.biomaterial_lineage.append(b)
@@ -159,7 +149,7 @@ class FileTransformer(Transformer):
         return "files"
 
     def create_documents(self, dss_bundle: DSSBundle) -> Sequence[ElasticSearchDocument]:
-        bundle = Bundle(dss_bundle)
+        bundle = api.Bundle(dss_bundle)
         for file in bundle.files.values():
             visitor = TransformerVisitor()
             # Visit the relatives of file
@@ -188,7 +178,7 @@ class SpecimenTransformer(Transformer):
         return "specimens"
 
     def create_documents(self, dss_bundle: DSSBundle) -> Sequence[ElasticSearchDocument]:
-        bundle = Bundle(dss_bundle)
+        bundle = api.Bundle(dss_bundle)
         for specimen in bundle.specimens:
             visitor = TransformerVisitor()
             # Visit the relatives of file
